==> analysis/01_data_wranging.R <==
#### Data wrangling. Author: Hirvin Azael Diaz Zepeda ####
library(readr)
library(dplyr)

# Import data base 14 - 11 - 2021
Cov_04_04 <- read_csv("data-raw/210404COVID19MEXICO.csv")

Cov <- Cov_04_04 %>% 
  select(2,6,10,11,12,13,14,16,33,36,40)

Cov <- Cov %>% 
  filter(CLASIFICACION_FINAL == 1 | CLASIFICACION_FINAL == 2 | CLASIFICACION_FINAL == 3) %>% 
  rename(ID = ID_REGISTRO,
         sex = SEXO,
         type = TIPO_PACIENTE,
         date_admission = FECHA_INGRESO,
         date_symptoms = FECHA_SINTOMAS,
         date_death = FECHA_DEF,
         intubated = INTUBADO,
         age = EDAD,
         icu = UCI,
         result = RESULTADO_LAB,
  ) %>% 
  filter(date_admission <= max(date_admission) - 7) %>% 
  mutate(state = ifelse(is.na(date_death), 0, 1))

save(Cov, file = "data/cov_04_04.Rdata")

colnames(QALEs_and_costs)

df_QALEs_cost <- QALEs_and_costs %>% 
  rename(QALE_m = Male...2,
         QALE_f = Female...3,
         Costs_m = Male...4,
         Costs_f = Female...5)

df_QALE <- df_QALEs_cost %>% 
  select(1,2,3) %>% 
  rename(male = QALE_m,
         female = QALE_f)

df_QALE <- gather(data = df_QALE, key = sex, value = QALE,-Edad)

save(df_QALE, file = "data/df_QALEs.Rdata")

df_Costs <- df_QALEs_cost %>% 
  select(1,4,5) %>% 
  rename(male = Costs_m,
         female = Costs_f)

df_Costs <- gather(data = df_Costs, key = sex, value = QALE,-Edad)

save(df_Costs, file = "data/df_Costs.Rdata")
